pyproject.toml
[project]
name = "Youtu-agent"
version = "0.1.0"
description = "Youtu-Agent is a simple but powerful framework for building, running, and evaluating agents."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.15",
    "akshare>=1.17.51",
    "art>=6.5",
    "arxiv>=2.2.0",
    "chunkr-ai>=0.1.0",
    "colorlog>=6.9.0",
    "docker>=7.1.0",
    "fpdf2>=2.8.4",
    "google-genai>=1.26.0",
    "gradio>=5.42.0",
    "hydra-core>=1.3.2",
    "ipython>=9.4.0",
    "jinja2>=3.1.6",
    "matplotlib>=3.10.3",
    "mcp>=1.12.3",
    # https://github.com/openai/openai-agents-python/issues/1395 | openai==1.97.1 & openai-agents==0.2.5
    "openai==1.99.6",
    "openai-agents==0.2.8",
    # https://github.com/Arize-ai/openinference/tree/main/python/instrumentation/openinference-instrumentation-openai
    "openinference-instrumentation-openai>=0.1.30",
    "openinference-instrumentation-openai-agents>=1.1.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pexpect>=4.9.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.4",
    "rich>=14.1.0",
    "seaborn>=0.13.2",
    "sqlmodel>=0.0.24",
    "tiktoken>=0.9.0",
    "toml>=0.10.2",
    "wikipedia-api>=0.8.1",
]

[project.optional-dependencies]
litllm = [
    "litellm>=1.77.0",
]

[dependency-groups]
dev = [
    "arize-phoenix-client>=1.14.1",
    "arize-phoenix-otel>=0.11.0",
    # datasets == 4.0.0 remove scripts, https://github.com/huggingface/datasets/issues/7693
    "datasets==3.6.0",
    "gpt-oss>=0.0.0",
    "gradio>=5.42.0",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.16",
    "mkdocstrings[python]>=0.30.0",
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.0.0",
    "pytest-timeout>=2.3.1",
    "pytest-xdist>=3.6.1",
    "pytest-mock>=3.14.0",
    "psutil>=6.1.0",
    "pillow>=11.3.0",
    "ruff>=0.12.8",
    # "tencentcloud-sdk-python>=3.0.1425",
    "tornado>=6.5.2",
]
documents = [
    "pymupdf>=1.26.3",
]
gaia = [
    "astor>=0.8.1",
    "docx2markdown>=0.1.1",
    "nest-asyncio>=1.6.0",
    "openpyxl>=3.1.5",
    "pillow>=11.3.0",
    "pydub>=0.25.1",
    "pymupdf>=1.26.3",
    "retry>=0.9.2",
    "tabulate>=0.9.0",
    "unstructured[docx,pptx]>=0.18.14",
    "validators>=0.35.0",
    "xls2xlsx>=0.2.0",
    "xmltodict>=0.15.0",
]
python-executor = [
    "chardet>=5.2.0",
    "matplotlib>=3.10.3",
    "pandas>=2.3.1",
    "scipy>=1.16.1",
    "seaborn>=0.13.2",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["utu"]


[tool.pytest.ini_options]
# pytest配置文件
pythonpath = "."
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"

# 测试路径
testpaths = ["tests"]

# 标记定义
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "performance: marks tests as performance tests",
    "edge_case: marks tests as edge case tests",
    "unit: marks tests as unit tests",
    "akshare: marks tests that require AKShare data",
    "chart: marks tests that generate charts",
    "report: marks tests that generate reports",
    "financial: marks tests related to financial analysis"
]

# 测试发现模式
python_files = [
    "test_*.py",
    "*_test.py"
]
python_classes = [
    "Test*"
]
python_functions = [
    "test_*"
]

# 日志配置
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(name)s: %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

# 基础选项
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "-v",
    "-s"
]

# 超时设置（需要pytest-timeout插件）
# timeout = 300
# timeout_method = "thread"

# 覆盖率设置（需要pytest-cov插件）
# addopts = ["--cov=utu", "--cov-report=term-missing"]

# 过滤警告
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore::ResourceWarning"
]

# 最小版本要求
minversion = "6.0"

[tool.uv]
index-url = "https://pypi.org/simple"
# extra-index-url = [
#     "https://mirrors.tencent.com/repository/pypi/tencent_pypi/simple",
# ]

[tool.ruff]
line-length = 120
target-version = "py312"
exclude = [
    "**/*.pyi",
    ".*",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
isort = { combine-as-imports = true, known-first-party = ["utu"] }

[tool.ruff.lint.pydocstyle]
convention = "google"


[tool.mypy]
strict = true
disallow_incomplete_defs = false
disallow_untyped_defs = false
disallow_untyped_calls = false

# [tool.pylint.messages_control]
# disable = ["broad-except", "bare-except", "invalid-name"]
